+++ lib/src/main/java/com/smileidentity/compose/SmileIDExt.kt
@@ -29,6 +29,8 @@ import com.smileidentity.util.randomUserId
 import com.smileidentity.viewmodel.document.DocumentVerificationViewModel
 import com.smileidentity.viewmodel.document.EnhancedDocumentVerificationViewModel
 import com.smileidentity.viewmodel.viewModelFactory
+import kotlinx.collections.immutable.ImmutableMap
+import kotlinx.collections.immutable.persistentMapOf
 import java.io.File
 import java.net.URL
 
@@ -49,6 +51,7 @@ import java.net.URL
  * @param showAttribution Whether to show the Smile ID attribution or not on the Instructions screen
  * @param showInstructions Whether to deactivate capture screen's instructions for
  * SmartSelfie.
+ * @param partnerParams Custom values specific to partners
  * @param colorScheme The color scheme to use for the UI. This is passed in so that we show a Smile
  * ID branded UI by default, but allow the user to override it if they want.
  * @param typography The typography to use for the UI. This is passed in so that we show a Smile ID
@@ -63,6 +66,7 @@ fun SmileID.SmartSelfieEnrollment(
     allowAgentMode: Boolean = false,
     showAttribution: Boolean = true,
     showInstructions: Boolean = true,
+    partnerParams: ImmutableMap<String, String> = persistentMapOf(),
     colorScheme: ColorScheme = SmileID.colorScheme,
     typography: Typography = SmileID.typography,
     onResult: SmileIDCallback<SmartSelfieResult> = {},
@@ -76,6 +80,7 @@ fun SmileID.SmartSelfieEnrollment(
             allowAgentMode = allowAgentMode,
             showAttribution = showAttribution,
             showInstructions = showInstructions,
+            partnerParams = partnerParams,
             onResult = onResult,
         )
     }
@@ -96,6 +101,7 @@ fun SmileID.SmartSelfieEnrollment(
  * displayed allowing toggling between the back camera and front camera. If not allowed, only the
  * front camera will be used.
  * @param showAttribution Whether to show the Smile ID attribution or not on the Instructions screen
+ * @param partnerParams Custom values specific to partners
  * @param colorScheme The color scheme to use for the UI. This is passed in so that we show a Smile
  * ID branded UI by default, but allow the user to override it if they want.
  * @param typography The typography to use for the UI. This is passed in so that we show a Smile ID
@@ -109,6 +115,7 @@ fun SmileID.SmartSelfieAuthentication(
     jobId: String = rememberSaveable { randomJobId() },
     allowAgentMode: Boolean = false,
     showAttribution: Boolean = true,
+    partnerParams: ImmutableMap<String, String> = persistentMapOf(),
     colorScheme: ColorScheme = SmileID.colorScheme,
     typography: Typography = SmileID.typography,
     onResult: SmileIDCallback<SmartSelfieResult> = {},
@@ -121,6 +128,7 @@ fun SmileID.SmartSelfieAuthentication(
             isEnroll = false,
             allowAgentMode = allowAgentMode,
             showAttribution = showAttribution,
+            partnerParams = partnerParams,
             onResult = onResult,
         )
     }
@@ -153,6 +161,7 @@ fun SmileID.SmartSelfieAuthentication(
  * @param allowGalleryUpload Whether to allow the user to upload images from their gallery or not
  * @param showInstructions Whether to deactivate capture screen's instructions for Document
  * Verification (NB! If instructions are disabled, gallery upload won't be possible)
+ * @param partnerParams Custom values specific to partners
  * @param colorScheme The color scheme to use for the UI. This is passed in so that we show a Smile
  * ID branded UI by default, but allow the user to override it if they want.
  * @param typography The typography to use for the UI. This is passed in so that we show a Smile ID
@@ -173,6 +182,7 @@ fun SmileID.DocumentVerification(
     allowAgentMode: Boolean = false,
     allowGalleryUpload: Boolean = false,
     showInstructions: Boolean = true,
+    partnerParams: ImmutableMap<String, String> = persistentMapOf(),
     colorScheme: ColorScheme = SmileID.colorScheme,
     typography: Typography = SmileID.typography,
     onResult: SmileIDCallback<DocumentVerificationResult> = {},
@@ -199,6 +209,7 @@ fun SmileID.DocumentVerification(
                         documentType = documentType,
                         captureBothSides = captureBothSides,
                         selfieFile = bypassSelfieCaptureWithFile,
+                        partnerParams = partnerParams,
                     )
                 },
             ),
@@ -232,6 +243,7 @@ fun SmileID.DocumentVerification(
  * @param allowGalleryUpload Whether to allow the user to upload images from their gallery or not
  * @param showInstructions Whether to deactivate capture screen's instructions for Document
  * Verification (NB! If instructions are disabled, gallery upload won't be possible)
+ * @param partnerParams Custom values specific to partners
  * @param colorScheme The color scheme to use for the UI. This is passed in so that we show a Smile
  * ID branded UI by default, but allow the user to override it if they want.
  * @param typography The typography to use for the UI. This is passed in so that we show a Smile ID
@@ -251,6 +263,7 @@ fun SmileID.EnhancedDocumentVerificationScreen(
     allowAgentMode: Boolean = false,
     allowGalleryUpload: Boolean = false,
     showInstructions: Boolean = true,
+    partnerParams: ImmutableMap<String, String> = persistentMapOf(),
     colorScheme: ColorScheme = SmileID.colorScheme,
     typography: Typography = SmileID.typography,
     onResult: SmileIDCallback<EnhancedDocumentVerificationResult> = {},
@@ -276,6 +289,7 @@ fun SmileID.EnhancedDocumentVerificationScreen(
                         countryCode = countryCode,
                         documentType = documentType,
                         captureBothSides = captureBothSides,
+                        partnerParams = partnerParams,
                     )
                 },
             ),
@@ -303,6 +317,7 @@ fun SmileID.EnhancedDocumentVerificationScreen(
  * allowing toggling between the back camera and front camera. If not allowed, only the front
  * camera will be used.
  * @param showAttribution Whether to show the Smile ID attribution or not on the Instructions screen
+ * @param partnerParams Custom values specific to partners
  * @param colorScheme The color scheme to use for the UI. This is passed in so that we show a Smile
  * ID branded UI by default, but allow the user to override it if they want.
  * @param typography The typography to use for the UI. This is passed in so that we show a Smile ID
@@ -321,6 +336,7 @@ fun SmileID.BiometricKYC(
     jobId: String = rememberSaveable { randomJobId() },
     allowAgentMode: Boolean = false,
     showAttribution: Boolean = true,
+    partnerParams: ImmutableMap<String, String> = persistentMapOf(),
     colorScheme: ColorScheme = SmileID.colorScheme,
     typography: Typography = SmileID.typography,
     onResult: SmileIDCallback<BiometricKycResult> = {},
@@ -337,6 +353,7 @@ fun SmileID.BiometricKYC(
             jobId = jobId,
             allowAgentMode = allowAgentMode,
             showAttribution = showAttribution,
+            partnerParams = partnerParams,
             onResult = onResult,
         )
     }