+++ lib/src/test/java/com/smileidentity/networking/PartnerParamsAdapterTest.kt
@@ -24,7 +24,11 @@ class PartnerParamsAdapterTest {
             userId = "userId",
             extras = mapOf("extra1" to "value1", "extra2" to "value2"),
         )
-        val kepMap = mapOf("jobId" to "job_id", "userId" to "user_id", "jobType" to "job_type")
+        val kepMap = mapOf(
+            "jobId" to "job_id",
+            "userId" to "user_id",
+            "jobType" to "job_type",
+        )
 
         // when
         val jsonString = adapter.toJson(partnerParams)