+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 10.0.0-beta13 (unreleased)
 
 ### Added
+- Added `extras` as optional params on all job types
+- Added `idAuthorityBypassPhoto` on Sandbox BiometricKYC jobs
 - Added `allowAgentMode` option on Document Verification and Enhanced Document Verification
 
 ### Fixed